+++ src/generic/adf_dev_driver_nativ.c
@@ -1,5 +1,5 @@
 /*
- * adf_nativ.c
+ * generic/adf_dev_driver_nativ.c
  *
  * $Id$
  *
@@ -24,7 +24,7 @@
 
 #include<stdlib.h>
 #include<string.h>
-#include"adf_nativ.h"
+#include "adf_dev_driver_nativ.h"
 #include"adf_err.h"
 #include "adf_env.h"
